Improve code for svm implementation: * Generalized handler class and moved it to its own file * Introduced model and result serialization
